ai/uxforai/semantickernel/requirements.txt
semantic-kernel>=0.9.6b1
python-dotenv>=1.0.1